build(examples): refactor main script into different files
