+++ projects/ngx-meta/examples/src/constants.ts
@@ -1,0 +1,4 @@
+import { join } from 'path'
+
+export const PACKAGE_JSON = 'package.json'
+export const ANGULAR_CLI_BINARY_PATH = join('node_modules', '.bin', 'ng')
