+++ projects/ngx-meta/examples/src/copy-app-dir-into-project.ts
@@ -1,0 +1,11 @@
+import { getAppsDir, Log } from './utils.js'
+import { basename, join } from 'path'
+import { cp } from 'fs/promises'
+
+export async function copyAppDirIntoProject(appDir: string) {
+  Log.step('Copying app from into project')
+  const appDirName = basename(appDir)
+  const destination = join(getAppsDir(), appDirName)
+  await cp(appDir, destination, { recursive: true })
+  return destination
+}
