+++ projects/ngx-meta/examples/src/generate-tmp-dir-and-register-cleanup-callback.ts
@@ -1,0 +1,51 @@
+import { execa, execaSync } from 'execa'
+import { Log } from './utils.js'
+
+export async function generateTmpDirAndRegisterCleanupCallback(
+  doCleanUp: boolean,
+): Promise<string> {
+  const tmpDir = (await execa('mktemp', ['-d'])).stdout
+  registerAbortAndExitCallback(
+    doCleanUp ? () => cleanUpTmpDir(tmpDir) : () => {},
+  )
+  Log.debug('Temporary dir: "%s"', tmpDir)
+  return tmpDir
+}
+
+// https://stackoverflow.com/a/14032965/3263250
+function registerAbortAndExitCallback(cleanUpCallback: () => void) {
+  //👇 Otherwise this prevents script from exiting
+  // process.stdin.resume() // so the program will not close instantly
+
+  process.on('exit', cleanUpCallback)
+  // catches ctrl+c event
+  process.on('SIGINT', () => {
+    Log.info('SIGINT received')
+    process.exit(2)
+  })
+  // catches "kill pid" (for example: nodemon restart)
+  process.on('SIGUSR1', () => {
+    Log.info('SIGUSR1 received')
+    process.exit(2)
+  })
+  process.on('SIGUSR2', () => {
+    Log.info('SIGUSR2 received')
+    process.exit(2)
+  })
+  // catches uncaught exceptions
+  process.on('uncaughtException', (error) => {
+    Log.error('Uncaught exception')
+    console.log(error)
+    process.exit(3)
+  })
+}
+
+// Must be sync as it's an exit handler
+// https://stackoverflow.com/a/14032965/3263250
+function cleanUpTmpDir(tmpDir: string) {
+  Log.step('Cleanup')
+  if (tmpDir != '' && tmpDir != '/') {
+    Log.step('Deleting temporary dir')
+    execaSync('rm', ['-rf', tmpDir])
+  }
+}
