+++ projects/ngx-meta/examples/src/update-app-config.ts
@@ -1,0 +1,6 @@
+import { Log } from './utils.js'
+
+export async function updateAppConfig(appDir: string) {
+  Log.step('Updating app config from template')
+  Log.info('TODO')
+}
